+++ src/types/types.ts
@@ -10,7 +10,7 @@
 
 export interface Message {
     action: 'toggleTranslation';
-    feature: 'titles' | 'isolatedPlayerTitles' | 'audio' | 'description' | 'isolatedPlayerDescriptionSearch' | 'subtitles';
+    feature: 'titles' | 'audio' | 'description' | 'subtitles';
     isEnabled: boolean;
 }
 
@@ -43,6 +43,7 @@ export interface ExtensionSettings {
     subtitlesTranslation: {
         enabled: boolean;
         language: string;
+        asrEnabled: boolean;
     };
     youtubeDataApi: {
         enabled: boolean;